fixed some connection exp stuff

+++ ConnectionExperiments/Client.h
@@ -5,10 +5,11 @@
 #ifndef CONNECTIONEXPERIMENTS_CLIENT_H
 #define CONNECTIONEXPERIMENTS_CLIENT_H
 
+#include <cstddef>
 
 class Client {
 public:
-    virtual void run_throughput_tests(int data_size) = 0;
+    virtual void run_throughput_tests(size_t data_size) = 0;
     virtual void run_latency_tests() = 0;
     virtual ~Client() = default;
 };
